feat: Audit Event Consumer 구성
1. Audit Event Processor 패키지 리소스/기능 중심 구성 업데이트
2. 멀티모듈 JPA 의존성 주입 설정
