--- audit-event-processor/src/main/java/com/junsik/audit/processor/AuditEventProcessorConfiguration.java
@@ -1,9 +1,0 @@
-package com.junsik.audit.processor;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@ComponentScan
-@Configuration
-public class AuditEventProcessorConfiguration {
-}
