+++ audit-event-processor/src/main/java/com/junsik/audit/processor/core/domain/enums/AuditFieldsEnum.java
@@ -1,0 +1,16 @@
+package com.junsik.audit.processor.core.domain.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum AuditFieldsEnum {
+	ID("id"),
+	CREATE_USER("createUser"),
+	UPDATE_USER("updateUser");
+
+	private final String fieldName;
+
+	AuditFieldsEnum(String fieldName) {
+		this.fieldName = fieldName;
+	}
+}
