+++ audit-event-processor/src/main/java/com/junsik/audit/processor/core/domain/enums/EventPublishType.java
@@ -1,0 +1,7 @@
+package com.junsik.audit.processor.core.domain.enums;
+
+public enum EventPublishType {
+	IMMEDIATELY,
+	RESERVE,
+	REPEAT
+}
