+++ audit-event-processor/src/main/java/com/junsik/audit/processor/core/domain/utils/AuditFieldExtractor.java
@@ -1,0 +1,6 @@
+package com.junsik.audit.processor.core.domain.utils;
+
+@FunctionalInterface
+public interface AuditFieldExtractor<T, E> {
+	void execute(T t, E e);
+}
