--- audit-event-processor/src/main/java/com/junsik/audit/processor/domain/AuditFieldExtractor.java
@@ -1,6 +1,0 @@
-package com.junsik.audit.processor.domain;
-
-@FunctionalInterface
-public interface AuditFieldExtractor<T, E> {
-	void execute(T t, E e);
-}
