--- audit-event-processor/src/main/java/com/junsik/audit/processor/domain/enums/AuditEventType.java
@@ -1,7 +1,0 @@
-package com.junsik.audit.processor.domain.enums;
-
-public enum AuditEventType {
-	CREATED,
-	UPDATED,
-	DELETED
-}
