--- audit-event-processor/src/main/java/com/junsik/audit/processor/domain/enums/AuditFieldsEnum.java
@@ -1,16 +1,0 @@
-package com.junsik.audit.processor.domain.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum AuditFieldsEnum {
-	ID("id"),
-	CREATE_USER("createUser"),
-	UPDATE_USER("updateUser");
-
-	private final String fieldName;
-
-	AuditFieldsEnum(String fieldName) {
-		this.fieldName = fieldName;
-	}
-}
