--- audit-event-processor/src/main/java/com/junsik/audit/processor/domain/enums/EventPublishType.java
@@ -1,7 +1,0 @@
-package com.junsik.audit.processor.domain.enums;
-
-public enum EventPublishType {
-	IMMEDIATELY,
-	RESERVE,
-	REPEAT
-}
