--- chat-event-consumer/build.gradle
@@ -1,5 +1,0 @@
-description = "chat-event-consumer module"
-
-dependencies {
-    api project(":audit-event-processor")
-}
