--- chat-event-consumer/src/main/resources/application.yml
@@ -1,4 +1,0 @@
-spring:
-  kafka:
-    consumer:
-      bootstrap-servers: localhost:9092,localhost:9093,localhost:9094
