--- chat-server/src/main/java/com/junsik/chat/adapter/web/request/SendMessageRequest.java
@@ -1,21 +1,0 @@
-package com.junsik.chat.adapter.web.request;
-
-import com.junsik.chat.domain.command.CreateMessageCommand;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class SendMessageRequest {
-	private Long senderId;
-	private Long receiverId;
-	private String content;
-
-	public CreateMessageCommand toCommand() {
-		return CreateMessageCommand.builder()
-				.senderId(this.senderId)
-				.receiverId(this.receiverId)
-				.content(this.content)
-				.build();
-	}
-}
