--- chat-server/src/main/java/com/junsik/chat/domain/command/CreateMessageCommand.java
@@ -1,14 +1,0 @@
-package com.junsik.chat.domain.command;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class CreateMessageCommand {
-	private Long senderId;
-	private Long receiverId;
-	private String content;
-}
