--- chat-server/src/main/java/com/junsik/chat/domain/service/CreateMessageService.java
@@ -1,29 +1,0 @@
-package com.junsik.chat.domain.service;
-
-import com.junsik.chat.adapter.persistence.entity.MessageEntity;
-import com.junsik.chat.adapter.persistence.repository.ChatRepository;
-import com.junsik.chat.domain.command.CreateMessageCommand;
-import com.junsik.chat.domain.model.Message;
-import com.junsik.chat.port.CreateMessageUseCase;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@RequiredArgsConstructor
-@Transactional
-@Service
-class CreateMessageService implements CreateMessageUseCase {
-
-	private final ChatRepository chatRepository;
-
-	@Override
-	public Message execute(CreateMessageCommand command) {
-		MessageEntity messageEntity =
-				MessageEntity.builder()
-						.receiverId(command.getReceiverId())
-						.content(command.getContent())
-						.build();
-
-		return Message.of(chatRepository.save(messageEntity));
-	}
-}
