--- chat-server/src/main/java/com/junsik/chat/port/CreateMessageUseCase.java
@@ -1,8 +1,0 @@
-package com.junsik.chat.port;
-
-import com.junsik.chat.domain.command.CreateMessageCommand;
-import com.junsik.chat.domain.model.Message;
-
-public interface CreateMessageUseCase {
-	Message execute(final CreateMessageCommand command);
-}
