+++ sample-application/src/main/java/com/junsik/sample/domain/command/CreateMessageCommand.java
@@ -1,0 +1,14 @@
+package com.junsik.sample.domain.command;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class CreateMessageCommand {
+	private Long senderId;
+	private Long receiverId;
+	private String content;
+}
