+++ sample-application/src/main/java/com/junsik/sample/port/CreateMessageUseCase.java
@@ -1,0 +1,8 @@
+package com.junsik.sample.port;
+
+import com.junsik.sample.domain.command.CreateMessageCommand;
+import com.junsik.sample.domain.model.Message;
+
+public interface CreateMessageUseCase {
+	Message execute(final CreateMessageCommand command);
+}
